add a clang-format file. cleanup

+++ src/timeseries.cpp
@@ -37,8 +37,8 @@ void nucmath::TimeSeries::add(double time, double value)
 
 size_t nucmath::TimeSeries::bin(double t)
 {
-    if(t >= hist.getLowestEdge() && t < hist.getLowestEdge() + hist.nBins()*hist.getBinWidth())
-        return floor((t-hist.getLowestEdge())/hist.getBinWidth());
+    if(t >= hist.getLowestEdge() && t < hist.getLowestEdge() + hist.nBins() * hist.getBinWidth())
+        return static_cast<size_t>(floor((t - hist.getLowestEdge()) / hist.getBinWidth()));
     else
         throw std::invalid_argument("TimeSeries::bin(double x): x is outside the data range.");
 }
@@ -49,14 +49,14 @@ std::pair<double, double> nucmath::TimeSeries::data(size_t bin) const
     {
         if(entriesPerBin.data(bin).second > 0.0)
         {
-            const double y = hist.data(bin).second/(entriesPerBin.data(bin).second);
+            const double y = hist.data(bin).second / (entriesPerBin.data(bin).second);
             return std::pair<double, double>(hist.data(bin).first, y);
         }
         else
             return std::pair<double, double>(hist.data(bin).first, 0);
     }
     else
-        throw std::out_of_range("nucmath::TimeSeries::data(std::size_t bin): bin="+std::to_string(bin));
+        throw std::out_of_range("nucmath::TimeSeries::data(std::size_t bin): bin=" + std::to_string(bin));
 }
 
 bool nucmath::TimeSeries::isChanged(bool leaveChanged)
@@ -97,10 +97,10 @@ double nucmath::TimeSeries::mean()
     double sum = 0.0;
     for(size_t i = 0; i < hist.nBins(); i++)
     {
-        sum += hist.data(i).second/(entriesPerBin.data(i).second);
+        sum += hist.data(i).second / (entriesPerBin.data(i).second);
     }
 
-    return sum/hist.nBins();
+    return sum / hist.nBins();
 }
 
 bool nucmath::TimeSeries::save(const std::string& path) const
@@ -112,13 +112,13 @@ bool nucmath::TimeSeries::save(const std::string& path) const
     {
         for(size_t i = 0; i < nBins(); i++)
         {
-            const auto& [x,y] = data(i);
+            const auto& [x, y] = data(i);
             stream << nucmath::to_string(x, 10) << "\t" << y << std::endl;
         }
     }
     else
     {
-        std::cout<< "Hist::save: can't open " << path << std::endl;
+        std::cout << "Hist::save: can't open " << path << std::endl;
         return false;
     }
